requirements.txt
databricks-sdk
streamlit
dlt[duckdb]